Move env-based data to dedicated directory
